+++ packages/storefront-next-dev/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add local dev support to MRT Data Layer [#1215](https://github.com/commerce-emu/storefront-next/pull/1215)
 - Clean up RSC/React Server Components references from documentation and comments ([#1363](https://github.com/commerce-emu/storefront-next/pull/1363))
+- Fix incorrect SSR sourcemaps for files sharing a basename in local dev debugger (@W-21175764)
 - Remove internal "odyssey" codename references from codebase
 - Add `sfnext scapi` CLI used to support generating and managing custom API clients in template projects (@W-21549425)
 - Fix `generate-cartridge` to scan `config-metadata/` for aspect type definitions (pdp.json, plp.json) after recent move from `src/` (@W-21875428)